[yejikim]style: file 분할 및 헤더 함수 선언 추가
