+++ src/free_memory.c
@@ -1,4 +1,3 @@
-#include "head.h"
 #include "execute.h"
 
 void	free_unit_pipe(t_unit_head *cmd_lst)
