+++ src/handle_error.c
@@ -1,4 +1,3 @@
-#include "head.h"
 #include "execute.h"
 
 int	handle_default_error(char *message)
