+++ src/main.c
@@ -1,4 +1,4 @@
-#include "head.h"
+#include "execute.h"
 
 int	main(void)
 {
